refactor(traits): generating documentation for addons
