+++ addons/zz_generated_doc.go
@@ -1,0 +1,1 @@
+package addons
