+++ cmd/util/doc-gen/main.go
@@ -21,6 +21,8 @@ import (
 	"github.com/apache/camel-k/cmd/util/doc-gen/generators"
 	"github.com/spf13/pflag"
 	"k8s.io/gengo/args"
+
+	_ "github.com/apache/camel-k/addons"
 )
 
 func main() {
