+++ docs/modules/ROOT/pages/traits/master.adoc
@@ -1,0 +1,32 @@
+= Master Trait
+
+// Start of autogenerated code - DO NOT EDIT! (description)
+The Master trait allows to configure the integration to automatically leverage Kubernetes resources for doing
+leader election and starting *master* routes only on certain instances.
+
+It's activated automatically when using the master endpoint in a route, e.g. `from("master:telegram:bots")...`.
+
+
+This trait is available in the following profiles: **Kubernetes, Knative, OpenShift**.
+
+// End of autogenerated code - DO NOT EDIT! (description)
+// Start of autogenerated code - DO NOT EDIT! (configuration)
+== Configuration
+
+Trait properties can be specified when running any integration with the CLI:
+```
+kamel run --trait master.[key]=[value] integration.groovy
+```
+The following configuration options are available:
+
+[cols="2,1,5a"]
+|===
+|Property | Type | Description
+
+| master.enabled
+| bool
+| Can be used to enable or disable a trait. All traits share this common property.
+
+|===
+
+// End of autogenerated code - DO NOT EDIT! (configuration)
